requirements.txt
flask==2.3.2
gunicorn==21.2.0

tensorflow-cpu==2.14.0

numpy==1.23.5
pillow==10.0.1
opencv-python==4.8.1.78
matplotlib==3.7.2
scikit-learn==1.3.0
tqdm==4.66.1
Werkzeug==2.3.7
